Remove note about 40M SourceMap Limit
Noone can point to where such a limit is actually defined if we indeed have one.

At least for the supported `sentry-cli` based SourceMap upload flow.
It is quite possible that the 40M limit refers to the legacy "single release file upload" endpoints which are deprecated and rarely used, and we would rather work towards removing those endpoints altogether.