feat(CODEOWNERS): Workaround for PR reviewers (#13662)
A while ago we made this change
https://github.com/getsentry/sentry-docs/pull/8868 so that we can have
codeowner review requirement enforced, but still have a reference on who
owns what.
This prevented the codeowners from becoming a blocker especially when we
have teams working across time zones.

But it became a bit inconvenient as the owners of files don't get tagged
in PR automatically and informed for review.
This PR is meant to fix that.

The new workflow file will be triggered on all PRs, identify all the
modified files in the PR, and cross match with all the commented lines
in CODEOWNERS files, if there's any match it will add the owners as
reviewer automatically.

Note: Approvals from reviewers added by this PR will not be required
before PR can be merged, this is more of an informational and making
sure proper teams are involved and informed on changes.