feat(develop): Add `page` context (#13203)
This adds a new context to develop docs:

The new `page` context should be used to add information about the
current page that an event originated on. It should be used for
web-based client SDKs instead of putting this information into
`event.request`.

Relay will need to extract data from this (if available) instead of
`event.request`.