Adding Copyable AI Rule Snippets to Docs (#13752)
Adding snippet copy functionality that users can copy to use in tools
like Cursor, Windsurf, or any LLM.

Modified the Expandable component to support a copy property that can
used to make snippets copyable in a user friendly way. Screenshot below.
Typically this would be a separate PR - but unfortunately the
functionality needed to be in place in order for this to work.

![CleanShot 2025-05-19 at 1  38
54@2x](https://github.com/user-attachments/assets/996cbc62-41d2-473b-b00e-c027d284dc8c)

Content can be copied and added to a local rules file in the editor and
referenced in configuration options - for example...
- "Implement logging using examples in the sentryrules.md" 
- " Using the example in sentryrules.md implement custom spans on my API
routes"

Per internal discussion, starting with high usage frameworks to start
basic and evaluate

## IS YOUR CHANGE URGENT?  

Help us prioritize incoming PRs by letting us know when the change needs
to go live.
- [x] None: Not urgent, can wait up to 1 week+

---------

Co-authored-by: getsantry[bot] <66042841+getsantry[bot]@users.noreply.github.com>