ref(browser): Add section about limitations of tracing with the loader (#13804)
## DESCRIBE YOUR PR

This PR adds a "Limitations of Tracing" section to the JS loader page
(analogously to the Limitations of Errors-Only Capturing). The problem
with tracing and using the loader is the async nature of the loader and
a couple of pitfalls related to this with the order of initializing the
SDK and auto instrumentation picking up signals (e.g. fetch calls).

I also added the `Sentry.init` call to the "Guarding SDK Function Calls"
section, to make it clear that Sentry.init should happen in
`window.sentryOnLoad` and not in `Sentry.onLoad` (naming wasn't on our
side in this case 😅)

#### Why can't the loader call `Sentry.onLoad` always after
`Sentry.init`?

This is because historically, users used `Sentry.onLoad` but there were
timing issues because the function had to be registered after the loader
script was evaluated (see
https://github.com/getsentry/sentry/pull/58574). So therefore, in order
for the loader's default `Sentry.init` **not** to be called, that
user-`Sentry.init` call needs to happen before --> `Sentry.onLoad` must
be called beforehand. Changing this could break peoples' Sentry (loader)
setup.

#### Why not provide a `Sentry.afterInit` hook?

This would further imply that users should delay important components of
their page, thus decreasing perceived performance. This is the very
thing why users should use the loader -> less performance overhead of
the SDK

Also, it would potentially couple user functionality to the Sentry SDK
to initialize. In many cases (ad blockers, script blockers), the SDK
can't init correctly, so that user behaviour would never be invoked.

## IS YOUR CHANGE URGENT?  

Help us prioritize incoming PRs by letting us know when the change needs
to go live.
- [ ] Urgent deadline (GA date, etc.): <!-- ENTER DATE HERE -->
- [ ] Other deadline: <!-- ENTER DATE HERE -->
- [x] None: Not urgent, can wait up to 1 week+

## SLA

- Teamwork makes the dream work, so please add a reviewer to your PRs.
- Please give the docs team up to 1 week to review your PR unless you've
added an urgent due date to it.
Thanks in advance for your help!

## PRE-MERGE CHECKLIST

*Make sure you've checked the following before merging your changes:*

- [ ] Checked Vercel preview for correctness, including links
- [ ] PR was reviewed and approved by any necessary SMEs (subject matter
experts)
- [ ] PR was reviewed and approved by a member of the [Sentry docs
team](https://github.com/orgs/getsentry/teams/docs)

## EXTRA RESOURCES

- [Sentry Docs contributor guide](https://docs.sentry.io/contributing/)