docs(api): Use correct HTTP method for specific issue (#13854)
<!-- Use this checklist to make sure your PR is ready for merge. You may
delete any sections you don't need. -->

## DESCRIBE YOUR PR
While going through the API docs page, I noticed that the example around
mutating issues uses the POST method, but earlier on this docs page, it
explicitly states that any endpoints whose purpose is to just update
should be a PUT. Moreover, the endpoints in this example are real and
there is no POST for them.

## IS YOUR CHANGE URGENT?  
No