docs(js): Create Remix manual quick start guide (#13928)
<!-- Use this checklist to make sure your PR is ready for merge. You may
delete any sections you don't need. -->

## DESCRIBE YOUR PR
Converted the Remix Manual Setup guide into a Quick Start guide.

I removed the "Profiling" onboarding option because it is not in the
wizard.
If we want to keep it in the quick start guide(s), we should create a
separate GH issue to add Profiling to both guides as a manual step
(updating the wizard installer itself would be best, though)

There's also an update in the wizard QS guide:
"Step 2: Server-Side Performance Monitoring" -> This is an important
step that is included in the manual setup but not in the wizard.
@chargome is inquiring about updating the wizard installer.

## IS YOUR CHANGE URGENT?  

Help us prioritize incoming PRs by letting us know when the change needs
to go live.
- [ ] Urgent deadline (GA date, etc.): <!-- ENTER DATE HERE -->
- [ ] Other deadline: <!-- ENTER DATE HERE -->
- [x] None: Not urgent, can wait up to 1 week+

## SLA

- Teamwork makes the dream work, so please add a reviewer to your PRs.
- Please give the docs team up to 1 week to review your PR unless you've
added an urgent due date to it.
Thanks in advance for your help!

## PRE-MERGE CHECKLIST

*Make sure you've checked the following before merging your changes:*

- [ ] Checked Vercel preview for correctness, including links
- [ ] PR was reviewed and approved by any necessary SMEs (subject matter
experts)
- [ ] PR was reviewed and approved by a member of the [Sentry docs
team](https://github.com/orgs/getsentry/teams/docs)

## EXTRA RESOURCES

- [Sentry Docs contributor guide](https://docs.sentry.io/contributing/)

---------

Co-authored-by: Charly Gomez <charly.gomez@sentry.io>