fix: Correct loader script sentence (#14228)
The doc says:


![image](https://github.com/user-attachments/assets/22500327-1662-4e19-9fff-29e1ee3f8cb5)

But when you create a JavaScript project and check the settings,
everything is actually turned off. We [recently changed the defaults so
all options start
disabled](https://github.com/getsentry/sentry/pull/86395). Before, when
they were all checked by default, the page would send a request on load
to update the script loader configuration — but now it doesn’t.
Basically, everything seems to be False (disabled) by default, so this
PR updates the docs to reflect that.