docs(nuxt): Add low version notice to all start guides (#14316)
## DESCRIBE YOUR PR

The expandable "Are you using Nuxt version < 3.14.0?" was only visible
on the "Manual Setup" docs. It's now added to the prerequisite section
to be included in all start guides.

## IS YOUR CHANGE URGENT?  

Help us prioritize incoming PRs by letting us know when the change needs
to go live.
- [ ] Urgent deadline (GA date, etc.): <!-- ENTER DATE HERE -->
- [ ] Other deadline: <!-- ENTER DATE HERE -->
- [ ] None: Not urgent, can wait up to 1 week+

## SLA

- Teamwork makes the dream work, so please add a reviewer to your PRs.
- Please give the docs team up to 1 week to review your PR unless you've
added an urgent due date to it.
Thanks in advance for your help!