ref: Make note about supported versions more generic (#14364)
We have an info note on some Python integration pages that suggests
switching to an older SDK major if a framework/Python version is not
supported on the current one. Currently it mentions specific SDK
versions. Make it more generic so that it doesn't get out of date if we
release a new major.