Nextjs source map doc updates (#14433)
## DESCRIBE YOUR PR
Updating the source map includes for nexjs to bring back the explainer
on why source maps are so important, and give a little bit of detail
around configuring.

Removing the nodejs blog post link from the javascript sourcemap doc.
It's outdated, and it's irrelevant to the other frameworks. We can
always add it into the node.js includes if we want it back, though it
should be updated if we want it back.

## IS YOUR CHANGE URGENT?  

Help us prioritize incoming PRs by letting us know when the change needs
to go live.
- [ ] Urgent deadline (GA date, etc.): <!-- ENTER DATE HERE -->
- [ ] Other deadline: <!-- ENTER DATE HERE -->
- [x] None: Not urgent, can wait up to 1 week+

## SLA

- Teamwork makes the dream work, so please add a reviewer to your PRs.
- Please give the docs team up to 1 week to review your PR unless you've
added an urgent due date to it.
Thanks in advance for your help!

## PRE-MERGE CHECKLIST

*Make sure you've checked the following before merging your changes:*

- [ ] Checked Vercel preview for correctness, including links
- [ ] PR was reviewed and approved by any necessary SMEs (subject matter
experts)
- [ ] PR was reviewed and approved by a member of the [Sentry docs
team](https://github.com/orgs/getsentry/teams/docs)

---------

Co-authored-by: Shannon Anahata <shannonanahata@gmail.com>
Co-authored-by: Sergiy Dybskiy <sergiy.dybskiy@sentry.io>