docs(self-hosted): update Linux compatibility with Amazon Linux 2023 (#14476)
Previously, it was stated that RHEL-based distro is a no-go for
installing self-hosted Sentry. But somebody did it with Amazon Linux
2023 (which is a Fedora derivative, very similar to RHEL). interestingly
with SELinux enabled, and... you can just read the screenshot below. I'm
adding that to the docs since that is so useful.

<img width="1168" height="905" alt="image"
src="https://github.com/user-attachments/assets/85dc9b25-d176-4166-8f98-1f56ca8d727b"
/>