docs(react-native): Update the docs in accordance to how the current structure of an Expo project looks like (#14545)
A relatively minor change: the current version of docs for Expo seems a
bit outdated in a sense that it asks you to make a change to the project
that doesn't match the actual project that's being generated by Expo.
This PR fixes that and makes it a bit easier for newcomers to understand
what changes need to be made.