docs(browser): Delete deprecated `profilesSampler` (#14549)
## DESCRIBE YOUR PR


Removes all occurrences of `profilesSampler` in the JS browser SDKs as
this API is not available and also deprecated, so we won't implement it.

-
https://github.com/getsentry/sentry-javascript/issues/16773#issuecomment-3143858577
- https://github.com/getsentry/sentry-javascript/issues/17279

## IS YOUR CHANGE URGENT?  

Help us prioritize incoming PRs by letting us know when the change needs
to go live.
- [ ] Urgent deadline (GA date, etc.): <!-- ENTER DATE HERE -->
- [ ] Other deadline: <!-- ENTER DATE HERE -->
- [ ] None: Not urgent, can wait up to 1 week+

## SLA

- Teamwork makes the dream work, so please add a reviewer to your PRs.
- Please give the docs team up to 1 week to review your PR unless you've
added an urgent due date to it.
Thanks in advance for your help!

## PRE-MERGE CHECKLIST

*Make sure you've checked the following before merging your changes:*

- [ ] Checked Vercel preview for correctness, including links
- [ ] PR was reviewed and approved by any necessary SMEs (subject matter
experts)
- [ ] PR was reviewed and approved by a member of the [Sentry docs
team](https://github.com/orgs/getsentry/teams/docs)