chore(issues): Update capture_message for Python
Using capture_message with a high-cardinality message is one of the best
ways to defeat our grouping algorithms and give yourself a _very_ noisy
issue stream.

Starting with Python for now so we can easily iterate on messaging here
but we will want to apply similar changes to all platforms in the
future.