feat: Update min js sdk version for logging and update console logging snippet (#14614)
In `9.41.0` of the JS SDK, we dropped the experimental `enableLogs`
option. Given we document snippets without the experimental option
everywhere, let's bump the min version in the docs:
https://github.com/getsentry/sentry-javascript/blob/develop/CHANGELOG.md#9410

This is the version we will lock in for GA.

While I was here, I also cleaned up the `consoleLoggingIntegration` so
that we follow the order log < warn < error.