feat(apple): Add session replay trouble shooting page with custom window handling (#13753)
## DESCRIBE YOUR PR

During debugging of the SDK we discovered that session replay for iOS
requires on the `window` property of the window scene delegate /
`UIWindowSceneDelegate` to be set. By default this will be set during
the bootstrapping of the app, but it appears that the property can also
be omitted and not configured.

This PR adds documentation to indicate to users how to deal with this.
It's also the start of having a troubleshooting section dedicated to
session replay:


https://sentry-docs-git-philprime-session-replay-scene-delegate.sentry.dev/platforms/apple/guides/ios/session-replay/troubleshooting/

## IS YOUR CHANGE URGENT?  

Help us prioritize incoming PRs by letting us know when the change needs
to go live.
- [ ] Urgent deadline (GA date, etc.): <!-- ENTER DATE HERE -->
- [ ] Other deadline: <!-- ENTER DATE HERE -->
- [X] None: Not urgent, can wait up to 1 week+

## SLA

- Teamwork makes the dream work, so please add a reviewer to your PRs.
- Please give the docs team up to 1 week to review your PR unless you've
added an urgent due date to it.
Thanks in advance for your help!

## PRE-MERGE CHECKLIST

*Make sure you've checked the following before merging your changes:*

- [X] Checked Vercel preview for correctness, including links
- [x] PR was reviewed and approved by any necessary SMEs (subject matter
experts)
- [ ] PR was reviewed and approved by a member of the [Sentry docs
team](https://github.com/orgs/getsentry/teams/docs)

## EXTRA RESOURCES

- [Apple
Documentation](https://developer.apple.com/documentation/uikit/uiwindowscenedelegate/window)

---------

Co-authored-by: Jasmin <77064737+jas-kas@users.noreply.github.com>