docs(devservices): Fill out all the `devservices up --modes=` possibilities (#14824)
I noticed that the mode `migrations` wasn't in the list, so i wanted to
add that.

But then i went and added everything else and threw them into some
categories that seemed reasonable. Idk if this is the best grouping, or
needed at all, but it's easy enough to throw the idea out there and then
delete it if it's too much.

included in the list in an added `replay` mode that depends on
https://github.com/getsentry/sentry/pull/98951