feat(logs): Add `sentry.replay_id` to logs spec (#14838)
resolves
https://linear.app/getsentry/issue/SDK-35/update-develop-docs-to-mention-sentryreplay-id-attribute

Attaching a replay id directly to logs makes it easy for users to go
from log -> replay. It also vastly simplifies querying for logs in the
frontend.

Right now we have only specced attaching replay to client-side logs, but
in the future we can add the functionality that sdks can parse baggage
to grab the `replay_id` and attach the id to logs.