feat(python): Make proxy options clearer (#14816)
## DESCRIBE YOUR PR
It's currently not documented how to make the Python SDK not use any
proxy. The available options for setting a proxy are confusing, with
`None` meaning we'll actually try to pick up proxy configuration from
environment variables. In order to not use a proxy, the options need to
be set to `""`. Documenting this.

---------

Co-authored-by: Alex Krawiec <alex.krawiec@sentry.io>