feat(node): Add new options to nativeNodeFetchIntegration (#15282)
<!-- Use this checklist to make sure your PR is ready for merge. You may
delete any sections you don't need. -->

## DESCRIBE YOUR PR

This is the documentation for
https://github.com/getsentry/sentry-javascript/pull/17996 (it also adds
the missing `spans`).

Now I think I did a little hack, since these options are not available
in `node-core`, but `javascript.node-core` is not a identifier. So I
show a little warning for only `javascript.node` that it is not
available in the `node-core` package. Since all other integrations use
`@sentry/node` anyways, we don't have to show this warning so that is
why I use the `<PlatformSection />` as well.

## IS YOUR CHANGE URGENT?  

Help us prioritize incoming PRs by letting us know when the change needs
to go live.
- [ ] Urgent deadline (GA date, etc.): <!-- ENTER DATE HERE -->
- [ ] Other deadline: <!-- ENTER DATE HERE -->
- [x] None: Not urgent, can wait up to 1 week+

## SLA

- Teamwork makes the dream work, so please add a reviewer to your PRs.
- Please give the docs team up to 1 week to review your PR unless you've
added an urgent due date to it.
Thanks in advance for your help!

## PRE-MERGE CHECKLIST

*Make sure you've checked the following before merging your changes:*

- [ ] Checked Vercel preview for correctness, including links
- [ ] PR was reviewed and approved by any necessary SMEs (subject matter
experts)
- [ ] PR was reviewed and approved by a member of the [Sentry docs
team](https://github.com/orgs/getsentry/teams/docs)

## EXTRA RESOURCES

- [Sentry Docs contributor guide](https://docs.sentry.io/contributing/)