Adding logs callout in relevant breadcrumbs docs. (#15432)
This change is not typically how we would want to move forward with
changes - being that we would normallly focus on making these changes on
each SDK.

Breadcrumbs are relatively low impact changes - and the content being
added is informational only. 

Making this change wider as an exception with the understanding we can revert specific areas if needed.