ref(develop/scopes): Remove attribute `type` as a user-provided property (#15522)
This is a proposal after working on scope attributes for JS. We don't
have to merge this but IMHO it makes sense.

This PR makes a slight change to the scope attributes specification:

When users want to set a scope attribute with a unit, the previous spec
currently required them to also pass `type` explicitly.

```ts
// previously
scope.setAttribute({
  simpleAttr: 'hi',
  attrWithUnit: {value: 120, type: 'integer', unit: 'ms'}
})

// now
scope.setAttribute({
  simpleAttr: 'hi',
  attrWithUnit: {value: 120, unit: 'ms'}, // serilized to integer attribute with value 120, unit ms
  attrWithoutUnit: {value: 120}, // serialized to integer attribute with value 120
  attrWithObject: {foo: 'bar'} // serialized to stringified string attribute
})
```

This PR changes to spec to preferrably remove the `type` property from
the attribute object in favour of letting the SDK internally check and
set the correct type. Two advantages:
1. Less overhead for users
2. No/less room for error for languages where we can't type strictly

This is a SHOULD, so if SDKs are constrained in some way, they can still
require the user to pass the type explicitly.

Furthermore, this PR also updates a recommendation to leave set
attribute format as-is, so that users can consume the attributes in
`beforeSendLog` and other callbacks, in the same structure as they are
initially set. This again is a SHOULD because most importantly, it must
be compatible with how attributes are consumed in said callbacks today
already.

---------

Co-authored-by: Michi Hoffmann <cleptric@users.noreply.github.com>