ref(sveltekit): Clarify limitations of auto instrumentation for Kit<2.31.0 (#15592)
Clarify that auto instrumentation will not work correctly if a SvelteKit
versions below 2.31.0 are used. This came up in
https://github.com/getsentry/sentry-javascript/issues/18247.

Note: I intentionally opted not to show the `--import` alternative I
mentioned in the issue because it would be a 3rd (already outdated) way
of setting up the SDK. Furthermore, success would vary greatly based on
the used Kit adapter and deployment platform.