feat: Re-organize log and trace drains page (#15642)
After doing a read through of
https://github.com/getsentry/sentry-docs/issues/15595, I realized we
could probably better incorporate details on forwarders, which I added
as part of https://github.com/getsentry/sentry-docs/pull/15623.

This hopefully makes this section a bit easier to grok, which is
important as I'm going to link to this doc from onboarding within a
couple of days, so it's going to get a lot more traffic.

Thank you @KyleTryon for kick starting the work on this, made it much
easier for me to conceptualize the needed sections on this.


https://sentry-docs-git-abhi-refactor-drains-index.sentry.dev/product/drains/