feat: Doc on introducing breaking changes to SDKs (#15647)
Add a page to help frame discussions about breaking changes in SDKs.
Discuss examples of breaking changes as well as changes which are not
apparently breaking, but we need to be mindful of them because they
significantly affect the user's Sentry experience. And finally, provide
some tips on making breaking changes in a way that introduces less
friction.


---------

Co-authored-by: Daniel Griesser <daniel.griesser.86@gmail.com>
Co-authored-by: Philipp Hofmann <philipp.hofmann@sentry.io>
Co-authored-by: Alex Krawiec <alex.krawiec@sentry.io>