docs(nextjs): Move webpack build config under `webpack` namespace (#15691)
We are moving the only-webpack related configurations under `webpack`
namespace to better fit our goal of communicating the SDK readiness for
Turbopack by default.

This involved moving several options to be nested under a `webpack` path
in the build configuration.

This is meant to be released once this
https://github.com/getsentry/sentry-javascript/pull/18343/ goes live, so
I will keep this open until then. Will keep it in draft as well.

The SDK PR adds warnings for those deprecated options and warns users
about using them if detected so I don't think we need to do that here
once it goes live.

---------

Co-authored-by: Sarah Mischinger <sarah@codingwriter.com>